Endian bugs, minor httpd fixes, relocations
. Fix endian bugs for MacOSX

. Fix tests for MacOSX ... another craptastic variant sprintf().

. Do 304 error return on if-modified-since for httpd

. Fix a couple of relocations, changing...
  const char *foo = "...";
  ...to...
  static const char foo[] = "...";

. Stop trying to build examples with mallinfo(), if it doesn't exist.

. Don't build ex_tdns ... as it doens't use Vstr anyway, and requires threads.

. Bump release number.

git-archimport-id: james@and.org--2004-code/vstr--main--1--patch-64

+++ scripts/coverage.sh
@@ -41,17 +41,22 @@ function del()
 
 function linkup()
 {
-  cd src
+  for dir in src examples; do
+  cd $dir
 
   if $doln; then
-   lndir ../$s/../src
+   lndir ../$s/../$dir
   fi
 
-  for i in .libs/*.da; do
-    ln -f $i; rm -f $i
-  done
+# Newer GCCs put them in the $srcdir
+  if [ ! -f ex_highlight-ex_highlight.da ]; then
+    for i in .libs/*.da; do
+      ln -f $i; rm -f $i
+    done
+  fi
 
   cd ..
+  done
 }
 
 function cov()
